+++ src/components/common/AddEntry/AddEntry.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import CategorySelector from '../CategorySelector';
 import { connect } from 'react-redux';
-import { addIncome, addOutcome } from '../../../redux/actions';
+import { addIncome, addOutcome } from '../../../redux/expensesManager/actions';
 import { getEntryModel, getEntryCategoryOption } from '../../../helpers/entriesHelper'; 
 
 import { withRouter } from 'react-router-dom';
